add sv_reftype_len() and make sv_reftype() be a wrapper for it
sv_reftype() mostly returns strings whose length is known at compile
time, so we can avoid a strlen() call if we return the length.

Additionally, the non-length interface is potentially buggy in the
face of class names which contain "\0", therefore providing a way
to obtain the true length allows us to avoid any trickyness.