Teach t/porting/podcheck.t that *_pm.PL files generate *.pm files.
Without this, it assumes that cpan/Devel-PPPort/PPPort_pm.PL and
lib/Devel/PPPort.pm have a duplicate =name. This can be covered up
by adding an exception for lib/Devel/PPPort.pm, but that hides the
underlying cause. The (believed) duplicate name was first exposed when
df80274d3278c640 made t/porting/podcheck.t case-insensitive, so as to
work on classic VMS filesystems, part of which required removing a skip
on all files matching /\.PL$/