fix broken optimisation of high code point unicode floating matches
If a unicode floating substr cannot be converted to a valid latin1 string
we end up with a non SvOK substr for latin1 matches, which we would pessimize
and allow the main engine to try the pattern. However we know in this case
we cannot match at all, so we can instead bail out right away. This also
prevents us doing anything illegal with the substr SV later.