fix off-by-one error in uninitialized warnings subscript finding logic
When we use an uninitialized var in a hash we try to find the var, so
we can show the subscript. An off by one error prevented us from seeing
the items in the last bucket.

We expect to see this:

   $ ./perl -Ilib -MDevel::Peek -wle'our %foo7=("foo"=>"bar","baz"=>undef); \
   $SIG{__WARN__}=sub { print STDERR @_; Dump(\%foo7); }; print sprintf "\n%s:%s %s:%s",%foo7;'

   Use of uninitialized value $foo7{"baz"} in sprintf at -e line 1.

But when baz was stored in the last bucket (1/8th of the time) it fails
to find it:

   $ ./perl -Ilib -MDevel::Peek -wle'our %foo7=("foo"=>"bar","baz"=>undef); \
   $SIG{__WARN__}=sub { print STDERR @_; Dump(\%foo7); }; print sprintf "\n%s:%s %s:%s",%foo7;'

   Use of uninitialized value within %foo7 in sprintf at -e line 1.

This patch fixes the off by one error. It does not correct any broken
tests it might cause. I will do that in a follow up patch.