rip out HvREHASH logic and replace it with random hash seed per process
The general idea of the HvREHASH() logic is to avoid hashing attacks by
noticing degenerate hash patterns and triggering the use of a random
seed.

However the attack is avoided if the attacker cannot predict the hash
ordering of the perl process they are talking to. By randomizing the
hash seed at perl process startup we make the ordering unpredictable
and remove the need for the HvREHASH logic and all of its overhead.