optimize S_minus_v
With the copyright lines, use more CPP cating. This reduced the number of
instructions, and removes a couple alignment null in .rdata image section.
UNDER_CE presents a challange because of the wce_hitreturn(); call.
BINARY_BUILD_NOTICE is arbitrary code. ActivePerl uses it and it contains
a perlio printf call. So that can not be CPP catted unfortunatly. If a
build is not UNDER_CE and BINARY_BUILD_NOTICE is undef, then the copyright
line is catted with the license line. LOCAL_PATCH_COUNT is a C expression,
like offsetof, which is not a preprocessor constant, so add
PERL_GIT_UNPUSHED_COMMITS_LEN and LOCAL_PATCH_COUNT_CONST which are pure
CPP constants. LOCAL_PATCH_COUNT can be changed to LOCAL_PATCH_COUNT_CONST
in the future.

Although S_minus_v is not hot and is rarely used, making the interp
image smaller, and therefore faster is a good idea. There should not be
any user visible changes to -v with this patch. switches.t contains a
regexp for -v already so no further tests were added.

This patch is part of #116296.