Make PERL_DEBUG_READONLY_COW and TRACK_MEMPOOL compatible
They both have headers at the beginning of the allocated memory.
PERL_TRACK_MEMPOOL is unique, in that it tries to free a pointer,
based not on what safesysmalloc returned, but on what it stored in
its own header.  So it needs to know about the extra COW header
when using pointer arithmetic to determine the pointer to pass
to safesysfree.