code following eval {} not always optimised
In something like this

    eval { 1 while 1 };
    $x = $a[0];

The optimising of the while loop makes Perl_rpeep() miss processing the
chain of ops from the OP_LEAVETRY onwards. So in the code above for
example, the alem wont be optimised into an alemfast.

Fix this by explicitly recursing into the entertry->op_other branch
(which actually points at the leavetry).

The infinite loop above can be broken by, for example, a signal handler
calling die.