When we share a refcounted_he_hek() we should mask off most of the flags
The refcounted_he flags and the normal HE flags overlap for the low
bits, but not the rest.

Perl_refcounted_he_new_pvn sets up hekflags with the flags required for
refcounted_he's. It then passes it into S_share_hek_flags(), which then
uses it in the comparisons for the lookup in PL_strtab.

This results in one entry per type, and since each entry has the same
hash value produces a collision "attack" on the bucket, which could
produce something like 6 * 2 collisions in a bucket worst case (although
there is no evidence of anything this extreme).

This appears to be unnecessary, as we store the hekflags in the
refcounted entry independently of the HEK.

Therefore we mask off the bits we care about, which dedupe the HEKs
in the string table. We still get collissions for WASUTF8 and
NON-WASUTF8, but for now it will do.