Add support for new warning categories outside of "all"
When someone suggests a new warning on p5p it always often up being
argued about on the basis that it'll break existing code, and that we
shouldn't add warnings for possibly legitimate code just because it's
unusual or odd.

As I pointed out in a discussion about RT #121025 (see [1]) we only keep
having this discussion because until now we've had no facility to add
new warnings outside of the default set that'll be retroactively enabled
for everything that does 'use warnings'. This patch introduces such a
facility.

To demonstrate this I'm adding a warning for something that was added as
a warning in the past, but pulled out because it was deemed too
controversial at the time: warning about the use of grep in void
context.

That warning was added back in v5.10.0-218-g74295f0 but quickly pulled
out in v5.10.0-230-gf5df478. See [2] for the discussion about it at the
time.

Now if you do:

    use warnings;
    grep /42/, (1,2);

You'll get no warnings as before, but if you do:

    use warnings qw(pedantic); # Or its sole subcategory: void_unusual
    grep /42/, (1,2);

You'll get a warning about "Unusual use of grep in void context". To
turn off this warning it's *not* sufficient to do:

    no warnings;

You need to do:

    no warnings qw(pedantic);

I'm willing to change that, but first we should ask ourselves whether
this should continue to remain a symmetric operation:

    {use,no} warnings ['all'];

This patch incorporates work from Andreas Guðmundsson
<andreasg@nasarde.org> who picked up an earlier version of mine and
figured out the change being made to mg.c here. That change removes an
optimization in the ${^WARNING_BITS} magic which might make things a tad
slower.

1. https://rt.perl.org/Ticket/Display.html?id=121025#txn-1276663
2. http://www.nntp.perl.org/group/perl.perl5.porters/2007/12/msg131922.html